+++ Changelog.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - macOS Monterey: There was a warning when creating a system report for the
   first time.
 ### Added
-- Export `render` and `rootReducer` as an object `testUtils` to aid testing.
+- Export `render`, `dispatchTo`, and `rootReducer` as an object `testUtils` to
+  aid testing.
 ### Changed
 - Move nrf-device-lib-js to dependencies as the exposed types system uses
   types from it.